lab4 --shift register triggered by clock and asynchronous reset
Determine the maximum and minimum values contained within the 4 registers comprising the shift register will be designed using clocked processes in VHDL. Two output registers will be used to store the maximum and minimum values which have been entered into the shift register file via the data input port.

The shift register file consists of 4 (R0, R1, R2, R3) registers with each register consisting of 4bits. These registers have a common reset signal. On every clock cycle, new data on the din port is stored into R0, and the old contents of R0 are shifted into R1, and the old contents of R1 are shifted into R2, etc. In this manner, the shift registers store the past 4 values which have been input.

A combinational logic block is used the determine the maximum and minimum values contained within the 4 shift registers. The output registers (max_reg and min_reg) are used to store the maximum and minimum values encountered during the operation of the circuit. These two registers have a load input (LD) which controls the selective loading of new data depending on whether the current values of the maximum and minimum signals (as produced by the max/min logic
block) are greater than or less than the current values stores in the two output registers. The values contained within the output registers are displayed on the LEDs of the FPGA board.

A 4-1 multiplexer is used to select any of the 4 registers within the shift register file to display the selected register value on the 4 remaining LEDS. A two-bit select input is used to select one of the 4 registers.