Create twos_complementer.vhd
this incompleted code was made to generate the two’s complement of a binary number: starting from the least significant bit position, retain all the 0’s and the first 1, after which complement each of the remaining bits.

An architecture for a two’s complementer consists of an 8-bit shift register with parallel load capability, a counter, and a control unit. The control unit generates the 7 required control signals : D, Shift, Load, Inc, Clr, ld_done, and clr_done.

The algorithm begins by first loading a number into the 8-bit shift register and clearing the counter and setting the flip-flop to 0. The least significant bit of the shift register is examined, and based upon its value the control unit then sets the D signal to the appropriate value and shifts it into the most significant position of the shift register. This
process is repeated until all 8 bits of the original number have been examined. At this point, the shift register will contain the two’s complement of the original number. A “Done” flip-flop is used to indicate that the algorithm is complete.